walnut/requirements.txt
django==2.1.3
SQLAlchemy==1.2.14
celery==4.2.1
jsonpath==0.80
ldap3==2.5.1
rc-django-celery-results==1.0.1
djangorestframework
django-cors-headers
Markdown==3.0.1
urllib3
requests